backend/requirements.txt
#### Denise 
django==3.2.9
djangorestframework==3.12.4
mysqlclient== 2.1.0
django-cors-headers==3.10.0
django-filter==21.1
django_mysql==4.3.0
